pyproject.toml
[project]
name = "hedgehog"
version = "1.0.0"
description = "Hierarchical Evaluation of Drug GEnerators tHrOugh riGorous filtration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.21.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "pyyaml>=5.4.0",
    "rdkit>=2022.3.0",
    "datamol>=0.8.0",
    "medchem>=0.1.0",
    "joblib>=1.3.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "syba",
    "jinja2>=3.0.0",
    "plotly>=5.0.0",
    "prolif>=2.1.0",
    "posecheck>=1.3.1",
    "shepherd-score>=1.2.1",
    "xgboost>=2",
    "posecheck-fast>=0.1.6 ; python_full_version >= '3.11'",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
lint = ["ruff", "mypy", "pyright"]
test = ["pytest", "pytest-cov"]

[project.scripts]
hedgehog = "hedgehog.main:app"
hedge = "hedgehog.main:app"


[tool.setuptools]
packages = {find = {where = ["src"]}}
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"*" = ["*.yml", "*.yaml", "*.csv", "*.json", "*.html"]
"hedgehog.vendor.moleval.metrics.NP_Score" = ["publicnp.model.gz"]
"hedgehog.vendor.moleval.metrics.SA_Score" = ["fpscores.pkl.gz"]
"hedgehog.vendor.moleval.metrics.fcd_torch" = ["ChemNet_v0.13_pretrained.pt"]

[tool.ruff]
target-version = "py310"
src = ["src"]
extend-exclude = [
    "modules",
    "src/hedgehog/vendor/**",
]
lint.select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
lint.ignore = [
    "COM812",  # conflicts with formatter
    "ISC001",  # conflicts with formatter
    "D100",    # module docstrings optional
    "D104",    # package docstrings optional
    "PT001",
    "PT023",
    "PLR0913", # allow >5 params
    "E501",    # allow line too long
    "B008",    # typer.Option in defaults - standard pattern
    "B905",    # zip() without strict= - ok for compatibility
]

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = [
    "F401",
    "F403",
]
"**/stages/structFilters/main.py" = ["F403", "F405"]
"**/stages/structFilters/utils.py" = ["E402"]
"**/stages/synthesis/main.py" = ["F403", "F405"]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::FutureWarning",
    "ignore::DeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src/hedgehog"]
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[tool.uv]
# posecheck 1.3.1 hard-pins pandas==2.0.0 which lacks wheels for Python 3.12+
# and fails to build from source (missing pkg_resources). Override to >=2.2.0.
override-dependencies = ["pandas>=2.2.0"]

[tool.uv.sources]
syba = { path = "modules/syba", editable = true }
